Move `CustomerSheet` data dependencies to a common module
