+++ paymentsheet/src/main/java/com/stripe/android/customersheet/injection/CustomerSheetDataCommonModule.kt
@@ -1,0 +1,72 @@
+package com.stripe.android.customersheet.injection
+
+import android.content.Context
+import com.stripe.android.BuildConfig
+import com.stripe.android.PaymentConfiguration
+import com.stripe.android.core.injection.ENABLE_LOGGING
+import com.stripe.android.core.injection.PUBLISHABLE_KEY
+import com.stripe.android.core.injection.STRIPE_ACCOUNT_ID
+import com.stripe.android.core.networking.AnalyticsRequestFactory
+import com.stripe.android.core.networking.NetworkTypeDetector
+import com.stripe.android.core.utils.ContextUtils.packageInfo
+import com.stripe.android.payments.core.analytics.ErrorReporter
+import com.stripe.android.payments.core.analytics.RealErrorReporter
+import com.stripe.android.payments.core.injection.PRODUCT_USAGE
+import dagger.Binds
+import dagger.Module
+import dagger.Provides
+import javax.inject.Named
+import javax.inject.Provider
+
+@Module
+internal interface CustomerSheetDataCommonModule {
+    @Binds
+    fun bindsErrorReporter(errorReporter: RealErrorReporter): ErrorReporter
+
+    companion object {
+        /**
+         * Provides a non-singleton PaymentConfiguration.
+         *
+         * Should be fetched only when it's needed, to allow client to set the publishableKey and
+         * stripeAccountId in PaymentConfiguration any time before presenting Customer Sheet.
+         *
+         * Should always be injected with [Lazy] or [Provider].
+         */
+        @Provides
+        fun providePaymentConfiguration(appContext: Context): PaymentConfiguration {
+            return PaymentConfiguration.getInstance(appContext)
+        }
+
+        @Provides
+        @Named(PUBLISHABLE_KEY)
+        fun providePublishableKey(
+            paymentConfiguration: Provider<PaymentConfiguration>
+        ): () -> String = { paymentConfiguration.get().publishableKey }
+
+        @Provides
+        @Named(STRIPE_ACCOUNT_ID)
+        fun provideStripeAccountId(
+            paymentConfiguration: Provider<PaymentConfiguration>
+        ): () -> String? = { paymentConfiguration.get().stripeAccountId }
+
+        @Provides
+        @Named(PRODUCT_USAGE)
+        fun providesProductUsage(): Set<String> = setOf("WalletMode")
+
+        @Provides
+        @Named(ENABLE_LOGGING)
+        fun providesEnableLogging(): Boolean = BuildConfig.DEBUG
+
+        @Provides
+        fun provideAnalyticsRequestFactory(
+            context: Context,
+            paymentConfiguration: Provider<PaymentConfiguration>
+        ): AnalyticsRequestFactory = AnalyticsRequestFactory(
+            packageManager = context.packageManager,
+            packageName = context.packageName.orEmpty(),
+            packageInfo = context.packageInfo,
+            publishableKeyProvider = { paymentConfiguration.get().publishableKey },
+            networkTypeProvider = NetworkTypeDetector(context)::invoke,
+        )
+    }
+}
