+++ paymentsheet/src/main/java/com/stripe/android/customersheet/injection/StripeCustomerAdapterComponent.kt
@@ -1,25 +1,14 @@
 package com.stripe.android.customersheet.injection
 
 import android.content.Context
-import com.stripe.android.BuildConfig
-import com.stripe.android.PaymentConfiguration
 import com.stripe.android.core.injection.CoreCommonModule
 import com.stripe.android.core.injection.CoroutineContextModule
-import com.stripe.android.core.injection.ENABLE_LOGGING
 import com.stripe.android.core.injection.IOContext
-import com.stripe.android.core.injection.PUBLISHABLE_KEY
-import com.stripe.android.core.injection.STRIPE_ACCOUNT_ID
-import com.stripe.android.core.networking.AnalyticsRequestFactory
-import com.stripe.android.core.networking.NetworkTypeDetector
-import com.stripe.android.core.utils.ContextUtils.packageInfo
 import com.stripe.android.customersheet.CustomerEphemeralKey
 import com.stripe.android.customersheet.CustomerEphemeralKeyProvider
 import com.stripe.android.customersheet.ExperimentalCustomerSheetApi
 import com.stripe.android.customersheet.SetupIntentClientSecretProvider
 import com.stripe.android.customersheet.StripeCustomerAdapter
-import com.stripe.android.payments.core.analytics.ErrorReporter
-import com.stripe.android.payments.core.analytics.RealErrorReporter
-import com.stripe.android.payments.core.injection.PRODUCT_USAGE
 import com.stripe.android.payments.core.injection.StripeRepositoryModule
 import com.stripe.android.paymentsheet.DefaultPrefsRepository
 import com.stripe.android.paymentsheet.PrefsRepository
@@ -31,8 +20,6 @@ import dagger.Component
 import dagger.Module
 import dagger.Provides
 import java.util.Calendar
-import javax.inject.Named
-import javax.inject.Provider
 import javax.inject.Singleton
 import kotlin.coroutines.CoroutineContext
 
@@ -40,6 +27,7 @@ import kotlin.coroutines.CoroutineContext
 @Component(
     modules = [
         StripeCustomerAdapterModule::class,
+        CustomerSheetDataCommonModule::class,
         StripeRepositoryModule::class,
         CoroutineContextModule::class,
         CoreCommonModule::class,
@@ -79,9 +67,6 @@ internal interface StripeCustomerAdapterModule {
     @Binds
     fun bindsCustomerRepository(repository: CustomerApiRepository): CustomerRepository
 
-    @Binds
-    fun bindsErrorReporter(errorReporter: RealErrorReporter): ErrorReporter
-
     companion object {
         @Provides
         fun provideTimeProvider(): () -> Long = {
@@ -99,50 +84,5 @@ internal interface StripeCustomerAdapterModule {
                 workContext
             )
         }
-
-        /**
-         * Provides a non-singleton PaymentConfiguration.
-         *
-         * Should be fetched only when it's needed, to allow client to set the publishableKey and
-         * stripeAccountId in PaymentConfiguration any time before presenting Customer Sheet.
-         *
-         * Should always be injected with [Lazy] or [Provider].
-         */
-        @Provides
-        fun providePaymentConfiguration(appContext: Context): PaymentConfiguration {
-            return PaymentConfiguration.getInstance(appContext)
-        }
-
-        @Provides
-        @Named(PUBLISHABLE_KEY)
-        fun providePublishableKey(
-            paymentConfiguration: Provider<PaymentConfiguration>
-        ): () -> String = { paymentConfiguration.get().publishableKey }
-
-        @Provides
-        @Named(STRIPE_ACCOUNT_ID)
-        fun provideStripeAccountId(
-            paymentConfiguration: Provider<PaymentConfiguration>
-        ): () -> String? = { paymentConfiguration.get().stripeAccountId }
-
-        @Provides
-        @Named(PRODUCT_USAGE)
-        fun providesProductUsage(): Set<String> = setOf("WalletMode")
-
-        @Provides
-        @Named(ENABLE_LOGGING)
-        fun providesEnableLogging(): Boolean = BuildConfig.DEBUG
-
-        @Provides
-        fun provideAnalyticsRequestFactory(
-            context: Context,
-            paymentConfiguration: Provider<PaymentConfiguration>
-        ): AnalyticsRequestFactory = AnalyticsRequestFactory(
-            packageManager = context.packageManager,
-            packageName = context.packageName.orEmpty(),
-            packageInfo = context.packageInfo,
-            publishableKeyProvider = { paymentConfiguration.get().publishableKey },
-            networkTypeProvider = NetworkTypeDetector(context)::invoke,
-        )
     }
 }
